Implement support for Representation Series in "Dimensionality Reduction" funcionts of representaion module
- Implement full support for Representation Series in "Dimensionality Reduction" functions of representation module
- add appropriate parameterized tests in `test_representation.py`

Note: the dimensionality reduction functions *do support* both flat and represenational input; this is because as discussed in #134 we will not implement a `to_repr` function, so users might have e.g. a flattened tfidf series they want to perform pca on -> they can use the function. Of course, we could remove the flat support (if we basically want to "forbid" users from using `VectorSeries` for tfidf/term_freq/count; we're not 100% sure.

Co-authored-by: Maximilian Krahn <maximilian.krahn@icloud.com>